+++ src/main/java/org/terasology/rendering/shader/ShaderParametersBlock.java
@@ -20,7 +20,7 @@ import static org.lwjgl.opengl.GL11.glBindTexture;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL13;
 import org.terasology.asset.Assets;
-import org.terasology.rendering.assets.GLSLShaderProgram;
+import org.terasology.rendering.assets.GLSLShaderProgramInstance;
 import org.terasology.rendering.assets.Texture;
 
 /**
@@ -34,7 +34,7 @@ public class ShaderParametersBlock extends ShaderParametersBase {
     }
 
     @Override
-    public void applyParameters(GLSLShaderProgram program) {
+    public void applyParameters(GLSLShaderProgramInstance program) {
         super.applyParameters(program);
 
         Texture terrainTex = Assets.getTexture("engine:terrain");
