+++ src/main/java/org/terasology/rendering/shader/ShaderParametersChunk.java
@@ -21,11 +21,9 @@ import org.terasology.asset.Assets;
 import org.terasology.config.Config;
 import org.terasology.editor.properties.Property;
 import org.terasology.game.CoreRegistry;
-import org.terasology.rendering.assets.GLSLShaderProgram;
+import org.terasology.rendering.assets.GLSLShaderProgramInstance;
 import org.terasology.rendering.renderingProcesses.DefaultRenderingProcess;
 import org.terasology.rendering.assets.Texture;
-import org.terasology.rendering.cameras.Camera;
-import org.terasology.rendering.world.WorldRenderer;
 
 import javax.vecmath.Vector4f;
 import java.util.List;
@@ -60,7 +58,7 @@ public class ShaderParametersChunk extends ShaderParametersBase {
     Property parallaxBias = new Property("parallaxBias", 0.05f, 0.0f, 0.5f);
     Property parallaxScale = new Property("parallaxScale", 0.05f, 0.0f, 0.5f);
 
-    public void applyParameters(GLSLShaderProgram program) {
+    public void applyParameters(GLSLShaderProgramInstance program) {
         super.applyParameters(program);
 
         Texture terrain = Assets.getTexture("engine:terrain");
