+++ src/main/java/org/terasology/rendering/shader/ShaderParametersPost.java
@@ -21,7 +21,7 @@ import org.terasology.componentSystem.controllers.LocalPlayerSystem;
 import org.terasology.config.Config;
 import org.terasology.editor.properties.Property;
 import org.terasology.game.CoreRegistry;
-import org.terasology.rendering.assets.GLSLShaderProgram;
+import org.terasology.rendering.assets.GLSLShaderProgramInstance;
 import org.terasology.rendering.renderingProcesses.DefaultRenderingProcess;
 import org.terasology.rendering.assets.Texture;
 import org.terasology.rendering.cameras.Camera;
@@ -50,7 +50,7 @@ public class ShaderParametersPost extends ShaderParametersBase {
     Property blurLength = new Property("blurLength", 0.15f, 0.0f, 1.0f);
 
     @Override
-    public void applyParameters(GLSLShaderProgram program) {
+    public void applyParameters(GLSLShaderProgramInstance program) {
         super.applyParameters(program);
 
         WorldRenderer worldRenderer = CoreRegistry.get(WorldRenderer.class);
